Expose delete_geff as a public function in core_io
