+++ packages/geff/src/geff/core_io/__init__.py
@@ -1,5 +1,11 @@
 from ._base_read import read_to_memory
 from ._base_write import write_arrays, write_dicts
-from ._utils import construct_var_len_props
+from ._utils import construct_var_len_props, delete_geff
 
-__all__ = ["construct_var_len_props", "read_to_memory", "write_arrays", "write_dicts"]
+__all__ = [
+    "construct_var_len_props",
+    "delete_geff",
+    "read_to_memory",
+    "write_arrays",
+    "write_dicts",
+]
