+++ packages/geff/src/geff/core_io/_base_write.py
@@ -1,19 +1,12 @@
 from __future__ import annotations
 
-import os
-import shutil
 import warnings
-from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
 
 import numpy as np
 
 from geff import _path
-from geff.core_io._utils import (
-    construct_var_len_props,
-    remove_tilde,
-    setup_zarr_group,
-)
+from geff.core_io._utils import construct_var_len_props, delete_geff, remove_tilde, setup_zarr_group
 from geff.validate.structure import validate_structure
 from geff_spec.utils import (
     add_or_update_props_metadata,
@@ -266,7 +259,7 @@ def write_arrays(
         try:
             validate_structure(geff_store)
         except ValueError as e:
-            _delete_invalid_geff(geff_store)
+            delete_geff(geff_store, zarr_format=zarr_format)
             raise ValueError(e.args[0] + "\nCannot write invalid geff.") from e
 
 
@@ -382,44 +375,3 @@ def write_props_arrays(
             prop_group[_path.DATA] = data
 
     return metadata
-
-
-def _delete_invalid_geff(store: StoreLike, zarr_format: Literal[2, 3] = 2) -> None:
-    """Delete an invalid geff after writing
-
-    Tries to handle multiple StoreLike inputs and avoids deleting non-geff contents
-    in the store
-
-    Args:
-        store (StoreLike): StoreLike geff that should be deleted
-        zarr_format (Literal[2, 3], optional): Zarr format used to write input store. Defaults to 2.
-    """
-    root = setup_zarr_group(store, zarr_format=zarr_format)
-
-    # Delete node and edge groups
-    del root[_path.NODES]
-    del root[_path.EDGES]
-
-    # If the root is empty, try to delete the root zarr
-    if len(list(root.keys())) == 0:
-        # Handle Path or str storelike
-        if isinstance(store, Path) or isinstance(store, str):
-            shutil.rmtree(store)
-        else:
-            # Try to get a valid path from the store
-            try:
-                path = store.path  # type: ignore
-                if os.path.exists(path):
-                    shutil.rmtree(path)
-            except AttributeError:
-                warnings.warn(
-                    "Cannot delete root zarr directory, but geff contents have been deleted",
-                    stacklevel=2,
-                )
-                del root.attrs["geff"]
-    else:
-        warnings.warn(
-            "Found non-geff members in zarr. Exiting without deleting root zarr.", stacklevel=2
-        )
-        # Delete geff metadata from attrs
-        del root.attrs["geff"]
